+++ reme4/components/file_parser/__init__.py
@@ -1,8 +1,8 @@
 """File parser components."""
 
-from .bare_file_parser import BareFileParser
 from .base_file_parser import BaseFileParser
+from .chunked_file_parser import ChunkedFileParser
 from .default_file_parser import DefaultFileParser
 from .linked_file_parser import LinkedFileParser
 
-__all__ = ["BareFileParser", "BaseFileParser", "DefaultFileParser", "LinkedFileParser"]
+__all__ = ["BaseFileParser", "ChunkedFileParser", "DefaultFileParser", "LinkedFileParser"]
